Introduce TruncPerfData and TruncText to utils package
Before we did not truncate performance_data or any other input to  mysql columns of data-type 'text' to length 64KB. This resulted in crashing of
icingadb, whenever input to one of these columns exceeded 64KB. Here, we add truncate.go to utils package. It contains TruncPerfData to truncate
performance_data and TruncText to truncate other inputs to mysql columns of data-type 'text' to required length.